Moved to play.core package
